move to latest uploads layout
